dev-requirements.txt
sqlalchemy==2.0.19
psycopg2-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests-oauthlib==1.3.1
